+++ baukis/app/controllers/application_controller.rb
@@ -7,13 +7,8 @@ class ApplicationController < ActionController::Base
   class Forbidden < ActionController::ActionControllerError; end
   class InAddressRejected < ActionController::ActionControllerError; end
 
-  rescue_from Exception, with: :rescue500
-  rescue_from Forbidden, with: :rescue403
-  rescue_from InAddressRejected, with: :rescue403
-  rescue_from ActionController::RoutingError, with: :rescue404
-  rescue_from ActiveRecord::RecordNotFound, with: :rescue404
+  include ErrorHandlers if Rails.env.production?
 
-  private
   def set_layout
     if params[:controller].match(%r{\A(staff|admin|customer)})
       Regexp.last_match[1]
@@ -22,18 +17,4 @@ class ApplicationController < ActionController::Base
     end
   end
 
-  def rescue500(e)
-    @exception = e
-    render 'errors/internal_server_error', status: 500
-  end
-
-  def rescue403(e)
-    @exception = e
-    render 'errors/forbidden', status: 403
-  end
-
-  def rescue404(e)
-    @exception = e
-    render 'errors/not_found', status: 403
-  end
 end
